cli/internal/logdetails: move to top-level "internal"
