+++ internal/logdetails/parse_logs.go
