+++ internal/logdetails/parse_logs_test.go
